Add the Net::HTTP adapter into its own gem
This extracts the Net::HTTP adapter into its own repository. The code is extracted from the main Faraday project and not altered. The idea behind this change is to create an ecosystem of different packages that you can use with Faraday, so that Faraday doesn’t have to ship with the packages itself.