Fix the build for Ruby version 2.4
Target version for Rubocop was 2.5 which removed the implicit begin rescue end block, however Faraday still supports Ruby 2.4, which means that we still have to use the begin rescue end block.

This change also update the Rubocop config to be in line with the Faraday repository minus the mandatory class documentation.