Version bump (3.0.0), set faraday dev dependency to 2.5 and required Ruby version to 2.6
Faraday 2.5 introduces a new streaming API and other changes that we want to take advantage on. However, releasing a new 2.x version of the gem would cause it to be pulled in projects where Faraday has not been updated to v2.5+ yet, causing errors.

For this reason, we're instead releasing this as v3.0, which only the next (compatible) version of faraday will allow to use.